+++ unit_tests/tests/errors.js
@@ -1,13 +1,12 @@
 const tap = require('tap');
 const { SECOND, wait } = require('../helpers/wait');
-const isProd = process.env.ISPROD === 'true';
-const Timer = require(isProd ? '../../time_class' : '../../index');
 
-tap.test('Errors feedBack tests', t => {
+const Timer = require('../timer');
 
+tap.test('Errors feedBack tests', (t) => {
   t.jobs = 11;
 
-  t.test('Callback is nor a Function nor a Promise', t => {
+  t.test('Callback is nor a Function nor a Promise', (t) => {
     const timer = new Timer(SECOND);
     try {
       timer.launchTimer({});
@@ -19,7 +18,7 @@ tap.test('Errors feedBack tests', t => {
     t.end();
   });
 
-  t.test('Callback is required', t => {
+  t.test('Callback is required', (t) => {
     const timer = new Timer(SECOND);
     try {
       timer.launchTimer();
@@ -31,7 +30,7 @@ tap.test('Errors feedBack tests', t => {
     t.end();
   });
 
-  t.test('Passed argument is not a Promise', async t => {
+  t.test('Passed argument is not a Promise', async (t) => {
     const timer = new Timer(SECOND);
     try {
       await timer.launchTimerPromise({});
@@ -43,7 +42,7 @@ tap.test('Errors feedBack tests', t => {
     t.end();
   });
 
-  t.test('Already Launch', t => {
+  t.test('Already Launch', (t) => {
     const timer = new Timer(SECOND);
     timer.launchTimer(() => {});
     try {
@@ -56,7 +55,7 @@ tap.test('Errors feedBack tests', t => {
     t.end();
   });
 
-  t.test('Already Launch (Promise)', async t => {
+  t.test('Already Launch (Promise)', async (t) => {
     const timer = new Timer(SECOND);
     const promise = wait(undefined, timer);
     try {
@@ -71,7 +70,7 @@ tap.test('Errors feedBack tests', t => {
     t.end();
   });
 
-  t.test('Destroy running timer', t => {
+  t.test('Destroy running timer', (t) => {
     const timer = new Timer(SECOND);
     timer.launchTimer(() => {});
     try {
@@ -84,7 +83,7 @@ tap.test('Errors feedBack tests', t => {
     t.end();
   });
 
-  t.test('Launching destroyed timer', t => {
+  t.test('Launching destroyed timer', (t) => {
     const timer = new Timer(SECOND);
     try {
       timer.destroy();
@@ -96,23 +95,26 @@ tap.test('Errors feedBack tests', t => {
         t.equal(msg, 'Timer is being deleted.');
       }
     } catch (e) {
-      t.fail('destroy() failed. Might not working properly.');  
+      t.fail('destroy() failed. Might not working properly.');
     }
     t.end();
   });
 
-  t.test('Create a new timer with existing ID (unforced)', async t => {
+  t.test('Create a new timer with existing ID (unforced)', async (t) => {
     try {
       new Timer(SECOND, { id: 1 });
       t.fail('Nothing went wrong...?');
     } catch (e) {
       const msg = e.message || e;
-      t.equal(msg, 'Timer already exist. To retrieve the existing one, please use `getById` Method.');
+      t.equal(
+        msg,
+        'Timer already exist. To retrieve the existing one, please use `getById` Method.'
+      );
     }
     t.end();
   });
 
-  t.test('Create a new timer with invalid ID (unforced)', async t => {
+  t.test('Create a new timer with invalid ID (unforced)', async (t) => {
     try {
       new Timer(SECOND, { id: true });
       t.fail('Nothing went wrong...?');
@@ -123,7 +125,7 @@ tap.test('Errors feedBack tests', t => {
     t.end();
   });
 
-  t.test('Create a new timer with invalid logger', async t => {
+  t.test('Create a new timer with invalid logger', async (t) => {
     try {
       new Timer(SECOND, { log: {} });
       t.fail('Nothing went wrong...?');