+++ unit_tests/tests/verbose.js
@@ -1,16 +1,15 @@
 const tap = require('tap');
 const { SECOND, wait } = require('../helpers/wait');
-const isProd = process.env.ISPROD === 'true';
-const Timer = require(isProd ? '../../time_class' : '../../index');
 
-const logs = ['A String', { log:'An object' }, ['An Array']]; 
+const Timer = require('../timer');
 
-tap.test('Verbose tests', t => {
+const logs = ['A String', { log: 'An object' }, ['An Array']];
 
+tap.test('Verbose tests', (t) => {
   t.jobs = 7;
 
   const testName0 = 'Launch with verbose (async)';
-  t.test(testName0, async t => {
+  t.test(testName0, async (t) => {
     const timer = new Timer(SECOND);
     await timer.launchTimer(wait(undefined, timer), null, testName0, ...logs);
     t.pass();
@@ -18,7 +17,7 @@ tap.test('Verbose tests', t => {
   });
 
   const testName1 = 'Launch with verbose (sync)';
-  t.test(testName1, async t => {
+  t.test(testName1, async (t) => {
     const timer = new Timer(SECOND);
     const promise = wait(undefined, timer);
     timer.launchTimer(() => {}, null, testName1, ...logs);
@@ -28,7 +27,7 @@ tap.test('Verbose tests', t => {
   });
 
   const testName2 = 'Update with verbose';
-  t.test(testName2, async t => {
+  t.test(testName2, async (t) => {
     const timer = new Timer(SECOND);
     await timer.launchTimer(wait(undefined, timer, testName2, ...logs));
     t.pass();
@@ -36,7 +35,7 @@ tap.test('Verbose tests', t => {
   });
 
   const testName3 = 'Done with verbose';
-  t.test(testName3, async t => {
+  t.test(testName3, async (t) => {
     const timer = new Timer(SECOND);
     timer.launchTimer(wait(undefined, timer));
     timer.done(testName3, ...logs);
@@ -45,7 +44,7 @@ tap.test('Verbose tests', t => {
   });
 
   const testName4 = 'Abort with verbose';
-  t.test(testName4, async t => {
+  t.test(testName4, async (t) => {
     try {
       const timer = new Timer(SECOND);
       const promise = timer.launchTimer(wait(undefined, timer));
@@ -58,32 +57,32 @@ tap.test('Verbose tests', t => {
   });
 
   const testName5 = 'Wrong Logger';
-  t.test(testName5, async t => {
+  t.test(testName5, async (t) => {
     try {
       const wrongLogger = (...args) => {
         throw args;
-      }
+      };
       const timer = new Timer(SECOND, { log: wrongLogger });
       const promise = timer.launchTimer(wait(undefined, timer));
       timer.abort(testName5, ...logs);
       await promise;
-    } catch(_) {
+    } catch (_) {
       t.pass();
     }
     t.end();
   });
 
   const testName6 = 'Custom Logger';
-  t.test(testName6, async t => {
+  t.test(testName6, async (t) => {
     try {
       const log = (...args) => {
-        console.log(...args.map(arg => `${arg}`.toUpperCase() ));
-      }
+        console.log(...args.map((arg) => `${arg}`.toUpperCase()));
+      };
       const timer = new Timer(SECOND, { log });
       const promise = timer.launchTimer(wait(undefined, timer));
       timer.abort(testName6, ...logs);
       await promise;
-    } catch(_) {
+    } catch (_) {
       t.pass();
     }
     t.end();