+++ CHANGELOG.md
@@ -3,6 +3,21 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.11] - 2022-06-24
+
+### Added
+
+- Code documentation using JSdoc syntaxe.
+
+### Changed
+
+- Readme.
+
+### Fixed
+
+- Wrong message error when passing Timer.prototype.launchTimer to a cron function (setTimeout, setInterval, Timer.prototype.launchTimer)
+- Crash when passing Timer prototypes (update, abort, done) to a cron function. Now, it ignores the call.
+
 ## [1.0.8] - 2021-03-19
 
 ### Added